Fixes #30 - Works around https://github.com/wangyoucao577/go-release-action/issues/143
The GitHub action we use is downloading the go source code to the
project's checked out clone. This is causing our `GitTreeState` to
always report as "dirty."

The file in question is always named `go-linux.tar.gz` (even on Windows
and Mac), since this is incorrect we can preemptively remove the tar
using a glob.