Remove JSON conditional check
Now that these messages print to `os.Stderr`, we do not need to
conditionally print them.

The conditional print was implemented to avoid breaking the JSON output
with non-JSON, but since it's going to `os.Stderr` now it doesn't
matter.

This opens the way for adding other output formats without needing to
maintain these conditionals.