Fixes 60 - `off` and `cycle`
`power off` and `power cycle` had `Args: cobra.MinimumNArgs(1)` set. This was previously removed for several other commands because it broke piping.

When piping occurs, at least with how `gru` accepts piped arguments, the `args[]` slice is empty. In order for piping to work, we can not set a minimum number of arguments.

There may be another way to handle piping, assuming there's a way to get in front of Cobra.